apps/drive-server/requirements.txt
numpy
Pillow
requests
boto3
websockets
pytest
pytest-asyncio
